Refactor: python, cts: Move RemoteRscFailure into the _cts module.
No code changes.
